consolidate into impl/response_format.rs
