+++ src/lang_model/impl/api/mod.rs
@@ -2,7 +2,6 @@ mod anthropic;
 mod chat_completion;
 mod gemini;
 mod openai;
-pub(super) mod schema_adapt;
 
 pub use anthropic::{AnthropicMarshal, AnthropicUnmarshal};
 pub use chat_completion::{ChatCompletionMarshal, ChatCompletionUnmarshal};
