+++ src/lang_model/impl/api/openai.rs
@@ -1,6 +1,6 @@
 use url::Url;
 
-use super::schema_adapt::SchemaAdapter as _;
+use super::super::response_format::ResponseSchemaMarshal;
 use crate::{
     datatype::Value,
     lang_model::{
@@ -28,11 +28,7 @@ impl LangModelProvider {
 #[derive(Clone, Debug, Default)]
 pub struct OpenAIMarshal;
 
-impl super::schema_adapt::SchemaAdapter for OpenAIMarshal {
-    fn adapt_schema(&self, schema: &crate::datatype::Value) -> crate::datatype::Value {
-        super::schema_adapt::for_strict_providers(schema)
-    }
-}
+impl ResponseSchemaMarshal for OpenAIMarshal {}
 
 fn marshal_message(msg: &Message, include_thinking: bool) -> Vec<Value> {
     let part_to_value = |part: &Part| -> Value {
@@ -221,7 +217,7 @@ impl Marshal<LangModelRequest<'_>> for OpenAIMarshal {
         }
         // top_k is not part of the OpenAI Responses spec; intentionally ignored.
         if let Some(ResponseFormat::JsonSchema(schema)) = req.response_format {
-            let wire_schema = self.adapt_schema(schema);
+            let wire_schema = self.marshal_response_schema(schema);
             body.as_object_mut().unwrap().insert(
                 "text".into(),
                 to_value!({
