+++ src/lang_model/impl/mod.rs
@@ -1,1 +1,2 @@
 pub(crate) mod api;
+pub(crate) mod response_format;
